Update postprocess_aweworker
Revised functionality to start multiple aweworkers on one VM.  With these changes, everything that's needed should all be in the config file for each AWE clientgroup, and in each config file directories are configured based on the group name.  The new start_service and stop_service scripts only call start_aweclient or stop_aweclient for each group.  The awec.cfg file is no longer used in this deployment strategy.  The PID file and startup logs are also moved out of the deployment directory (to accomodate sharing the deployment directory among multiple AWE worker VMs).

The hacks of the *_aweclient scripts are awful; ideally the changes should go to the AWE code, but I am too lazy to do that coding right now.