configure postprocess_narrative properly
I think there was a recent change to the narrative src/config.json file which sets the "ci" stanza of that file as the default.  These changes substitute the name of the deployment, and are tested in -next.  Since our production deploys are named proda and prodb, I also have a line that replaces either of those (or anything with "prod" in it) with prod (untested).